fix: better error reporting from file converter

+++ utils/raw_to_h5.sh
@@ -17,16 +17,7 @@ mzml_file="${raw_file%.raw}.mzML"
 h5_file="${raw_file%.raw}.h5"
 progress_file="${raw_file%.raw}.progress"
 failure_file="${raw_file%.raw}.failed"
-
-if [ -f "$h5_file" ]; then
-    >&2 echo "INFO: .h5 file found for ${raw_file} - skipping conversion."
-   exit 0
-fi
-
-if [ -f "$failure_file" ]; then
-    >&2 echo "INFO: .failed file found for ${raw_file} - skipping conversion."
-   exit 0
-fi
+done="false"
 
 function finish {
   if [ "$done" = "true" ]; then
@@ -37,28 +28,46 @@ function finish {
 }
 trap finish EXIT
 
-done="false"
-
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 time_stamper="${script_dir}/ts.py"
+
+if [ -f "$h5_file" ]; then
+   printf "INFO: .h5 file found for %s - skipping conversion.\n" "$raw_file" \
+	"${time_stamper}" | \
+	tee -a "${progress_file}"
+   exit 0
+fi
+
+if [ -f "$failure_file" ]; then
+   printf "INFO: .failed file found for %s - skipping conversion.\n" "$raw_file" \
+	"${time_stamper}" | \
+	tee -a "${progress_file}"
+   exit 0
+fi
 
 shifter "--env=PYTHONPATH=/src" "--image=doejgi/metatlas_shifter:latest" \
 	"${script_dir}/validate_file_name.py" "$raw_file" 2>&1 | \
 	"${time_stamper}" | \
 	tee "${progress_file}"
 
-# ThermoRawFileParser.sh should return non-zero exit code on error, but it doesn't
-# https://github.com/compomics/ThermoRawFileParser/issues/140
-# But the mzML to h5 conversion will fail nicely if the mzML files does not exist
 shifter "--image=${raw_image}" ThermoRawFileParser.sh \
 	"-i=${raw_file}" "-o=$(dirname "$raw_file")" -f=1 2>&1 | \
 	sed 's%^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} \([A-Z]+\)%\1, ThermoRawFileParser:%' | \
 	"${time_stamper}" | \
 	tee -a "${progress_file}"
 
+# ThermoRawFileParser.sh should return non-zero exit code on error, but it doesn't
+# https://github.com/compomics/ThermoRawFileParser/issues/140
+if [ ! -f "${mzml_file}" ] || [ ! -s "${mzml_file}" ]; then
+   printf "ERROR: raw to mzml conversion failed for %s, but return code was still 0.\n" "$raw_file" \
+	"${time_stamper}" | \
+	tee -a "${progress_file}"
+   exit 1
+fi
+
 shifter "--env=PYTHONPATH=/src" "--image=doejgi/metatlas_shifter:latest" \
-	"${script_dir}/mzml_to_h5.py" "${mzml_file}" | \
+	"${script_dir}/mzml_to_h5.py" "${mzml_file}" 2>&1 | \
 	"${time_stamper}" | \
 	tee -a "${progress_file}"
